// src/types/integration-menu.ts
export type TIntegrationMenuSlug =
  | "novu-inbox"
  | "sendgrid"
  | "ses"
  | "postmark"
  | "resend"
  | "brevo"
  | "mailgun"
  | "mailjet"
  | "twilio"
  | "plivo"
  | "aws-sns"
  | "vonage"
  | "sms77"
  | "telnyx"
  | "termii"
  | "fcm"
  | "apns"
  | "expo-push"
  | "onesignal"
  | "pushpad"
  | "pusher-beams"
  | "push-webhook"
  | "slack"
  | "discord"
  | "ms-teams"
  | "mattermost"
  | "whatsapp"
  | "zulip"
  | "react-email"
  | "vue-email"
  | "mjml"
  | "maizzle"
  | "brail"
  | "langchain"
  | "vercel-ai-sdk"
  | "launchdarkly"
  | "flagsmith"
  | "posthog"

export type TIntegrationMenuIcon = `integration-${TIntegrationMenuSlug}`